requirements.txt
Pillow
PySocks
PyOpenSSL
cffi
idna
kiwisolver
cryptography
pycparser
requests
charset-normalizer
numpy
torchvision
packaging
brotlipy
tornado
lxml
fonttools
urllib3
torch
pyparsing
wheel
matplotlib
six
setuptools
typing_extensions
munkres
certifi
python-dateutil
unicodedata2
pip
cycler
kiwisolver